:refactor: recycle: 에러 코드 및 프로젝트 구조 통일, 접근 제어자 수정
