+++ src/main/java/com/whereyouad/WhereYouAd/global/security/oauth2/service/CustomOAuth2UserService.java
@@ -6,8 +6,8 @@ import com.whereyouad.WhereYouAd.domains.user.exception.UserSignUpException;
 import com.whereyouad.WhereYouAd.domains.user.exception.code.UserErrorCode;
 import com.whereyouad.WhereYouAd.domains.user.persistence.entity.AuthProviderAccount;
 import com.whereyouad.WhereYouAd.domains.user.persistence.entity.User;
-import com.whereyouad.WhereYouAd.domains.user.persistence.mapper.AuthAccountConverter;
-import com.whereyouad.WhereYouAd.domains.user.persistence.mapper.UserConverter;
+import com.whereyouad.WhereYouAd.domains.user.application.mapper.AuthAccountConverter;
+import com.whereyouad.WhereYouAd.domains.user.application.mapper.UserConverter;
 import com.whereyouad.WhereYouAd.domains.user.persistence.repository.AuthProviderAccountRepository;
 import com.whereyouad.WhereYouAd.domains.user.persistence.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
